chore: move from Codec to CodecV2
This should be protocol backward compatible change (and breaking change in the source code) which should speed-up the process of gRPC
data encoding / decoding. Also unify our CodecV2 usage.

For https://github.com/siderolabs/talos/issues/9404
