+++ cmd/integration-test/main.go
@@ -6,10 +6,15 @@
 // Package main provides the entrypoint for the omni-integration-test binary.
 package main
 
-import "os"
+import (
+	"os"
+
+	_ "github.com/siderolabs/omni/cmd/acompat" // this package should always be imported first for init->set env to work
+	"github.com/siderolabs/omni/cmd/integration-test/pkg"
+)
 
 func main() {
-	if err := rootCmd.Execute(); err != nil {
+	if err := pkg.RootCmd().Execute(); err != nil {
 		os.Exit(1)
 	}
 }
